refactor: split diplomacy_relation_lookup into focused modules (#3825)
Extract relation constants and upsert primitives into sibling files while
keeping the public API unchanged via barrel re-exports.
