+++ packages/colonizethis_diplomacy/lib/colonizethis_diplomacy.dart
@@ -9,7 +9,9 @@ export 'src/diplomacy/diplomacy_event_logging.dart';
 export 'src/diplomacy/diplomacy_logging.dart';
 export 'src/diplomacy/diplomacy_phase_result.dart';
 export 'src/diplomacy/diplomacy_power_score.dart';
+export 'src/diplomacy/diplomacy_relation_constants.dart';
 export 'src/diplomacy/diplomacy_relation_lookup.dart';
+export 'src/diplomacy/diplomacy_relation_upsert.dart';
 export 'src/diplomacy/diplomacy_relation_updates.dart';
 export 'src/diplomacy/diplomacy_resolver.dart';
 export 'src/diplomacy/diplomacy_shared_helpers.dart';
